+++ popup/image-generator.html
@@ -29,7 +29,10 @@
       </div>
     </div>
   </div>
-
+  <script src="imagegen/DataLoader.js"></script>
+  <script src="imagegen/CanvasRenderer.js"></script>
+  <script src="imagegen/ImageProcessor.js"></script>
+  <script src="imagegen/OutputManager.js"></script>
   <script src="image-generator.js"></script>
 </body>
 </html>
